Bluetooth: Mesh: DFU17867 errata fix del_all
Fixes errata change for delete and delete all message handling in the
DFD server model.

Original problem statement:

"
Consider this scenario: a distribution is completed successfully, and
there are nodes in the Receivers List. Then a Firmware Delete All is
received.

The Distributor goes to Idle as per the current text, but it still has
a bunch of nodes in its Receivers List.

If the Distribution Client sends a Receivers Get, it will get nodes
that are complete. It can start a distribution with a list of nodes
that are complete.

This breaks a lot of things, because in that case, the setting of the
initial values for the nodes that are still in the Receivers List
(BLOB Transfer progress, Retrieved Update Phase etc.) is now
circumvented. Not to say that it's pretty confusing for a user to see
that you're idle, but the nodes list says it's complete (or failed).

There is another related problem:
Distribution is complete with FW List Index 0. There is an image with
Index 0 and Index 1. We receive Distribution Start with the exact same
parameters. We respond with “already done with that”. We receive FW
Delete By Index with Index 0. We receive Distribution Start with the
exact same parameters. We respond with “already done” - even though now
FW List Index 0 is a different firmware.

The most straightforward way to remedy both problems would be that
images can only be deleted when the Distribution Phase is Idle.
Then the procedure for removing unnecessary images or starting a new
distribution would always start with a Distribution Cancel or Receivers
Delete All - explicitly telling the Distribution Server that it does not
need to retain any old state anymore. All states will have known values,
and the Firmware List indices shifting will not be a problem.
"

Signed-off-by: Anders Storrø <anders.storro@nordicsemi.no>