Bluetooth: Mesh: Fix SegAck reception in SAR TX
This commit fixes 2 issues.

1. Table 3.20 of MeshPRFd1.1r15 states what valid SegAck is. One of the
   conditions is:

For the SeqAuth derived from the SeqZero field of the message, there is
at least one unacknowledged segment that the AckedSegments field of the
message reports as delivered.

Which means that only new SegAck should reset SAR Retransmissions Timer.

2. When transmitting segments, tx->seg_o is between 1 and tx->seg_n + 1.
Once transmission of all (remaining) segments are done, tx->seg_o is set
to 0. If new SegAck is received while the transport layer is not yet done
with transmissions of the remaining segments (which is also the case
when tx->seg_o > tx->seg_n), rescheduling tx->retransmit timer will brake
that procedure. This is because the tx-retransmit timer will also be
called from seg_sent callback. Considering current implementation, the
only way to handle this correctly is to prevent new SegAck message to
schedule the timer if the segments transmissions is in progress. In
other words, until SAR Retransmissions Timer is started.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>