Bluetooth: Mesh: Reschedule SAR Ack Timer on every new segment
According to new SAR logic, we need to restart SAR Acknowledgment
Timer on every new segment received. From MshPRFd1.1r16, section
3.5.3.4:

When the Processing Result is Next Segment and the destination is a
unicast address, the lower transport layer shall store the received
segment, shall start the SAR Acknowledgment Timer and the SAR Discard
Timer from the initial values, and shall set the AckedSegments value
for the SeqAuth derived from this message to indicate that the segment
with offset SegO has been received.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>