doc: bluetooth: mesh: Update DFU guide and samples documentation
After adding distributor and target samples, and implementing a shell
commands for DFD server to replace the Initiator, it become necessary to
align the DFU evaluation guide with the new shell commands. This lets to
reduce the distributor sample documentation by moving generic parts to
this guide. Some other parts of the DFU and BLOB guides are also updated
to clarify some aspects of the DFU procedure.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>