Bluetooth: Mesh: Fix sending Firmware Update Start message to multicast
Firmware Update Start message has Update Firmware Image Index field
which is unique for each Updating node since it is an index of the
Firmware Information List state of an Updating node. However, if the
index is the same for all Updating node, the message can be sent to a
multicast address.

This commit fixes transmission of Firmware Update Start message in
multicast mode.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>