Bluetooth: Mesh: Move blob_cli_broadcast_ctx to bt_mesh_blob_cli
Even though blob_cli_broadcast_ctx is initialized a `static const` so
that they are stored in flash, there is no need in having so many
contexts since they can be initialized "in place" and the actual context
storage could be moved to BLOB client. This would let us save some space
in flash. This also let us to get rid of some ugly code where context
configuration depends on mutlicast and unicast address, and on pull and
push modes.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>